Local Storage
-when the user adds a game through the form the game is added to the archive array, then a local storage item is created from the archive array.

Display:
when the archive is displayed a variable is set equal to "JSON.parse(localStorage.getItem('archive'));" and then the archive array is set equal to that variable.

Remove Game:
To remove a game - first, the archive is updated, then the local storage is completely cleared. I then create another local storage item set to the new archive array